Some fixes to make the configuration changes work for the app
